Conflicts:
	openair1/PHY/MODULATION/nr_modulation.h
	openair1/PHY/NR_TRANSPORT/nr_dlsch.c
	openair1/SCHED_NR/phy_procedures_nr_gNB.c

+++ openair1/PHY/MODULATION/nr_modulation.c
@@ -24,7 +24,7 @@
 extern short nr_mod_table[NR_MOD_TABLE_SIZE_SHORT];
 
 void nr_modulation(uint32_t *in,
-                   uint16_t length,
+                   uint32_t length,
                    uint16_t mod_order,
                    int16_t *out)
 {
