+++ openair1/PHY/MODULATION/nr_modulation.h
@@ -37,7 +37,7 @@
 */
 
 void nr_modulation(uint32_t *in,
-                   uint16_t length,
+                   uint32_t length,
                    uint16_t mod_order,
                    int16_t *out);
 
@@ -92,7 +92,6 @@ int nr_slot_fep_ul(NR_DL_FRAME_PARMS *frame_parms,
 \param Msc_PUSCH number of allocated data subcarriers
 */
 void nr_dft(int32_t *z,int32_t *d, uint32_t Msc_PUSCH);
-
 
 int nr_beam_precoding(int32_t **txdataF,
 	              int32_t **txdataF_BF,
